# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "지원 API",
    "version": "0.0.1",
    "description": "일반부원 및 개발부원 지원을 위한 API입니다."
  },
  "paths": {
    "/data1": {
      "post": {
        "summary": "일반부원 모집",
        "description": "일반부원 지원 정보를 등록합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "이름",
                    "example": "이제민"
                  },
                  "student_id": {
                    "type": "string",
                    "description": "학번",
                    "example": "2020440102"
                  },
                  "major": {
                    "type": "string",
                    "description": "전공",
                    "example": "전자전기컴퓨터공학부"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "지원자의 전화번호",
                    "example": "01096761452"
                  },
                  "motivation": {
                    "type": "string",
                    "description": "지원 동기",
                    "example": "동아리에 참여하고 싶습니다."
                  }
                },
                "required": ["name", "student_id", "major", "phone_number", "motivation"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "저장 완료, FE 측에서 메세지 전송",
            "content": {
              "application/json": {
                "example": "저장 완료"
              }
            }
          },
          "400": {
            "description": "잘못된 요청, BE 측에서 메세지 전송",
            "content": {
              "application/json": {
                "example": "${value}의 값이 누락되었습니다 / ${value}의 형식이 잘못 되었습니다"
              }
            }
          },
          "409": {
            "description": "이미 신청하셨습니다, FE 측 메세지 전송",
            "content": {
              "application/json": {
                "example": "중복된 신청입니다"
              }
            }
          },
          "500": {
            "description": "서버 에러, FE 측 메세지 전송",
            "content": {
              "application/json": {
                "example": "서버 에러"
              }
            }
          }
        }
      }
    },
    "/data2": {
      "post": {
        "summary": "개발부원 모집",
        "description": "개발부원 지원 정보를 등록합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "이름",
                    "example": "이제민"
                  },
                  "student_id": {
                    "type": "string",
                    "description": "학번",
                    "example": "2020440102"
                  },
                  "major": {
                    "type": "string",
                    "description": "전공",
                    "example": "전자전기컴퓨터공학부"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "전화번호",
                    "example": "01096761452"
                  },
                  "department": {
                    "type": "string",
                    "description": "지원 분야, design or frontend or backend",
                    "example": "design"
                  },
                  "motivation": {
                    "type": "string",
                    "description": "지원 동기",
                    "example": "동아리에 참여하고 싶습니다."
                  },
                  "project_description": {
                    "type": "string",
                    "description": "프로젝트 설명",
                    "example": "프로젝트 설명입니다."
                  },
                  "portfolio_pdf": {
                    "type": "string",
                    "format": "binary",
                    "description": "포트폴리오 PDF 파일"
                  },
                  "github_profile": {
                    "type": "string",
                    "description": "Github 프로필 URL",
                    "example": "https://github.com/username"
                  },
                  "github_email": {
                    "type": "string",
                    "description": "Github 이메일",
                    "example": "example@github.com"
                  },
                  "slack_email": {
                    "type": "string",
                    "description": "Slack 이메일",
                    "example": "example@slack.com"
                  },
                  "willing_general_member": {
                    "type": "boolean",
                    "description": "일반부원 지원 의사",
                    "example": "1"
                  }
                },
                "required": ["name", "student_id", "major", "phone_number", "department","motivation", "portfolio_pdf", "willing_general_member"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "저장 완료, FE 측에서 메세지 전송",
            "content": {
              "application/json": {
                "example": "저장 완료"
              }
            }
          },
          "400": {
            "description": "잘못된 요청, BE 측에서 메세지 전송",
            "content": {
              "application/json": {
                "example": "${value}의 값이 누락되었습니다 / ${value}의 형식이 잘못 되었습니다"
              }
            }
          },
          "409": {
            "description": "이미 신청하셨습니다, FE 측 메세지 전송",
            "content": {
              "application/json": {
                "example": "중복된 신청입니다"
              }
            }
          },
          "500": {
            "description": "서버 에러, FE 측 메세지 전송",
            "content": {
              "application/json": {
                "example": "서버 에러"
              }
            }
          }
        }
      }
    },
    "/event/can_participation": {
      "get": {
        "summary": "이벤트 참여 가능 여부 확인",
        "description": "DB 내 경품 수량을 확인해, 이벤트 참여 가능 여부 확인",
        "responses": {
          "200": {
            "description": "이벤트 참여 가능",
            "content": {
              "application/json": {
                "example": "이벤트 참여 가능"
              }
            }
          },
          "403": {
            "description": "이벤트 참여 불가",
            "content": {
              "application/json": {
                "example": "이벤트 참여 불가"
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "example": "server error"
              }
            }
          }
        }
      }
    },
    "/event/participation": {
      "post": {
        "summary": "이벤트 참여",
        "description": "이벤트 참여",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "참여자 이름",
                    "example": "이제민"
                  },
                  "student_id": {
                    "type": "string",
                    "description": "참여자 학번",
                    "example": "2020440102"
                  }
                },
                "required": ["name", "student_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "이벤트 로직 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "참여자 이름"
                    },
                    "student_id": {
                      "type": "string",
                      "description": "참여자 학번"
                    },
                    "winning": {
                      "type": "integer",
                      "description": "당첨 여부 1 or 0"
                    },
                    "goods": {
                      "type": "string",
                      "description": "당첨 상품"
                    }
                  },
                  "example": {
                    "name": "이제민",
                    "student_id": "2020440102",
                    "winning": 1,
                    "goods": "메가 커피"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "example": "서버 에러"
              }
            }
          }
        }
      }
    }
  }
}